cleanup - hide to_two_pubkeys behind verify_from_commit
